fix validations in products module

+++ src/app/components/admin/products/products.component.ts
@@ -4,6 +4,8 @@ import { MatPaginator } from '@angular/material/paginator';
 import { MatTableDataSource } from '@angular/material/table';
 import { ProductsService } from '../../services/products.service';
 import { EditProductComponent } from './edit-product/edit-product.component';
+import swal from 'sweetalert2';
+import { NotificationsService } from '../../services/notificaciones.service';
 
 @Component({
   selector: 'app-products',
@@ -11,7 +13,7 @@ import { EditProductComponent } from './edit-product/edit-product.component';
   styleUrls: ['./products.component.css']
 })
 export class ProductsComponent implements OnInit {
-  products: any =  [];
+  products: any = [];
   displayedColumns: string[] = ['title', 'category', 'description', 'brand', 'img', 'actions'];
   dataSource: any;
   @ViewChild(MatPaginator, { static: false })
@@ -19,7 +21,8 @@ export class ProductsComponent implements OnInit {
 
   constructor(
     private productsService: ProductsService,
-    public dialog: MatDialog
+    public dialog: MatDialog,
+    private notificationsService: NotificationsService
   ) { }
 
   ngOnInit() {
@@ -27,35 +30,60 @@ export class ProductsComponent implements OnInit {
   }
 
   getProducts() {
-    this.productsService.getAll().then(x=> {
+    this.productsService.getAll().then(x => {
       this.products = x.products;
       this.dataSource = new MatTableDataSource<any>(this.products);
       this.dataSource.paginator = this.paginator;
     }).catch()
   }
 
-  getProductById(id) {
-    this.productsService.getById(id).then(x=> {
-      if(x.id) {
-        
+  getProductById(id, index) {
+    this.productsService.getById(id).then(x => {
+      if (x.id) {
         const dialogRef = this.dialog.open(EditProductComponent, {
           disableClose: true,
           width: '25vw',
           height: 'auto',
-          data: { product: x}
+          data: { product: x }
         });
         dialogRef.afterClosed().subscribe(result => {
           if (result) {
+            this.products[index] = result;
+            this.dataSource = new MatTableDataSource<any>(this.products);
+            this.dataSource.paginator = this.paginator;
+            this.notificationsService.showSuccess("Producto modificado exitosamente");
           }
         })
       }
-      
     }).catch()
   }
 
-  deleteProduct(id) {
-    this.productsService.delete(id).then(x=> {
-    }).catch()
+  deleteProduct(id, index) {
+    const msgTitle = "Eliminar";
+    const msgText = "¿Seguro que desea elminar el producto?";
+    const confirmMessage = "Si, eliminar";
+    const cancelMessage = "Cancelar";
+    swal.fire({
+      title: msgTitle,
+      text: `${msgText}`,
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: confirmMessage,
+      cancelButtonText: cancelMessage,
+    }).then((result) => {
+      if (result.value) {
+        this.productsService.delete(id).then(x => {
+          if (x.id) {
+            this.products.splice(index, 1);
+            this.dataSource = new MatTableDataSource<any>(this.products);
+            this.dataSource.paginator = this.paginator;
+            this.notificationsService.showSuccess("Producto eliminado exitosamente");
+          }
+        }).catch()
+      }
+    })
   }
 
 }
